requirements.txt
azure-identity==1.15.0
openai==1.3.6
python-dotenv==1.0.0